#1521 - Created the SharedInfoField
